Restructure G4 tests: per-task files, remove concurrency/parity tests
Address review feedback:
- Split tests from 2 monolithic files into 5 per-task files matching
  repo convention (GivenA{TaskName}.cs)
- Remove concurrency tests: these pure in-memory tasks have no static
  state or process-level side effects, so static analysis is sufficient
- Remove output parity tests: redundant with behavioral tests that
  assert against known expected values
- Remove GivenTasksUseAbsolutePaths.cs: these tasks do no file I/O,
  so CWD-shifting tests add no value
- Keep all edge case tests (null, empty, empty ItemSpec, path preservation)
- Keep all behavioral tests (detect duplicates, grouping, error paths)
